MODDCB-76 Covering with unit tests
